Adding logic to check if services are running before attempting to stop them

+++ scripts/ubuntu/pentaho/install-9.0-sp.sh
@@ -21,8 +21,22 @@ else
 fi
 
 # https://stackoverflow.com/questions/3510673/find-and-kill-a-process-in-one-line-using-bash-and-regex/3510850#3510850
-kill -9 $(ps aux | grep '[p]ostgres.bin' | awk '{print $2}')
-kill -9 $(ps aux | grep '[t]omcat' | awk '{print $2}')
+
+postgresCheck=$(ps aux | grep '[p]ostgres.bin')
+
+if [ -z "$postgresCheck" ]; then
+  echo "postgres is not running"
+else
+  kill -9 $(ps aux | grep '[p]ostgres.bin' | awk '{print $2}')
+fi
+
+tomcatCheck=$(ps aux | grep '[t]omcat')
+
+if [ -z "$tomcatCheck" ]; then
+  echo "tomcat is not running"
+else
+  kill -9 $(ps aux | grep '[t]omcat' | awk '{print $2}')
+fi
 
 echo "Patching pentaho-server with $PATCH_VERSION"
 $INSTALLERS_BASE/PentahoServer-SP-$PATCH_VERSION.bin -i silent -DEULA_ACCEPT=true -DUSER_INSTALL_DIR=$PENTAHO_TO_PATCH/server/pentaho-server -DSILENT_LOG=$LOG_BASE/pentaho-server.log
